% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{create_rd}
\alias{create_rd}
\title{Creates R help entry}
\usage{
create_rd(spec_path)
}
\arguments{
\item{spec_path}{The file location of the YAML spec translation file.  It is a required argument, cannot be left empty.}
}
\description{
It uses the YAML spec file to create a single help entry.  Its output is a text vector.
}
\examples{
library(datalang)
my_spec <- system.file("specs/thisweek.yml", package = "datalang")
create_rd(my_spec)
}
