% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{datalang_help}
\alias{datalang_help}
\title{Wrapper for 'help()' function}
\usage{
datalang_help(topic)
}
\arguments{
\item{topic}{A quoted or unquoted name of the data set or function}
}
\description{
Prevents unnecessary rd files to be shipped with the "host" package.  It builds and displays the data set
help, but it does it in the RStudio Viewer pane, as oppossed to the Help pane.  If the requested topic does
not match to thel help entries available via 'datalang', then 'datalang_help()' will forward the topic to
the regular 'help()' function.
}
