% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{datalang_help_add}
\alias{datalang_help_add}
\title{Adds help entry to the current R session}
\usage{
datalang_help_add(obj, spec_path, package = NULL)
}
\arguments{
\item{obj}{The object's name to be tracked}

\item{spec_path}{The location of the YAML spec file}

\item{package}{Name of the "host" package. Optional.}
}
\description{
It is a mechanism to track which data sets, or functions, to display via the Help pane,
and which to build and display on demand.
}
