% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{folder_data}
\alias{folder_data}
\title{Translates and saves multiple a data sets}
\usage{
folder_data(spec_folder = "inst/specs", data_folder = "data")
}
\arguments{
\item{spec_folder}{The path to the folder where the YAML spec files are located. Defaults to 'inst/specs'.}

\item{data_folder}{The target folder location where the 'rda' file will be save to.}
}
\description{
Cycles through all of the spec files inside a folder specified in the 'spec_folder' argument.
The function translates and saves them to a specified folder.

Saves a translated versions of the data set in 'rda' format.  It wraps the translate_data()
function, and saves the output.

This function is meant for packages that will ship with a copy of the translated data set.
}
\examples{
library(datalang)
my_spec_folder <- system.file("specs", package = "datalang")
folder_data(my_spec_folder, tempdir())
}
