% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{folder_rd}
\alias{folder_rd}
\title{Creates rd files for multiple data sets}
\usage{
folder_rd(spec_folder = "inst/specs", rd_folder = "man")
}
\arguments{
\item{spec_folder}{The path to the folder where the YAML spec files are located. Defaults to 'inst/specs'.}

\item{rd_folder}{The target folder location where the 'rd' file will be save to. Defaults to 'man'.}
}
\description{
Cycles through all of the spec files inside a folder specified in the 'spec_folder' argument.
The function translates and saves them to a specified folder.

Saves the files to the folder specified in the 'rd_folder' argument.

This function is meant for packages that will ship with a copy of the translated data set.
}
\examples{
library(datalang)
my_spec_folder <- system.file("specs", package = "datalang")
folder_rd(my_spec_folder, tempdir())
}
