% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_folder_functions}
\alias{load_folder_functions}
\title{Translates and loads multiple a funciton into an R environment}
\usage{
load_folder_functions(
  spec_folder = "inst/specs",
  verbose = FALSE,
  envir = globalenv(),
  package = NULL
)
}
\arguments{
\item{spec_folder}{The path to the folder where the YAML spec files are located. Defaults to 'inst/specs'.}

\item{verbose}{If TRUE, it returns a list object with specs that where processed}

\item{envir}{The target environment where the translated data set will be loaded to. Defaults to the base R environment.}

\item{package}{Name of the package as a character variable. It is used in the help tracker.}
}
\description{
Cycles through all of the spec files inside a folder specified in the 'spec_folder' argument.
The function translates and loads each new data set to a specified R environment.

This function is meant for packages that will do not wish to ship a copy of the data set, and
wish to execute the translation on the fly, preferably when the "host" package is attached.
}
\examples{
library(datalang)
my_spec_folder <- system.file("specs", package = "datalang")
load_folder_data(my_spec_folder)
}
