% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{load_translation}
\alias{load_translation}
\title{Translates and loads a data set into an R environment}
\usage{
load_translation(spec_path, envir = baseenv(), package = NULL)
}
\arguments{
\item{spec_path}{The file location of the YAML spec translation file.  It is a required argument, cannot be left empty.}

\item{envir}{The target environment where the translated data set will be loaded to. Defaults to the base R environment.}
}
\description{
Translates and loads the translation to a specified R environment.

This function is meant for packages that will do not wish to ship a copy of the data set, and
wish to execute the translation on the fly, preferably when the "host" package is attached.
}
\examples{
library(datalang)
my_spec <- system.file("specs/thisweek.yml", package = "datalang")
load_translation(my_spec)
}
