% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{save_rd}
\alias{save_rd}
\title{Creates and saved R help file}
\usage{
save_rd(spec_path = NULL, rd_folder = "man")
}
\arguments{
\item{spec_path}{The file location of the YAML spec translation file.  It is a required argument, cannot be left NULL.}

\item{rd_folder}{The target folder location where the 'rd' file will be save to. Defaults to 'man'.}
}
\description{
Saves a the help output, from 'create_rd()', into an 'rd' file with the same name as the translated dataset.

This function is meant for packages that will ship with a copy of the translated data set.
}
\examples{
library(datalang)
my_spec <- system.file("specs/thisweek.yml", package = "datalang")
save_rd(my_spec, tempdir())
}
