% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thisweek}
\alias{thisweek}
\title{Sample data set}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
thisweek
}
\description{
Three days of morning and afternoon temperature measurements
}
\details{
\describe{
  \item{day}{Day of the week}
  \item{morning}{Temperature recorded at 10 AM}
  \item{afternoon}{Temperature recorded at 3 PM}
}
}
\keyword{datasets}
