% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_data}
\alias{translate_data}
\title{Translates a data set}
\usage{
translate_data(spec_path, .data = NULL)
}
\arguments{
\item{spec_path}{The file location of the YAML spec translation file.  It is a required argument, cannot be left empty.}

\item{.data}{A tibble or data.frame object that overrides the one specified in the spec file. Defaults to NULL.}
}
\description{
This is the main translation function in 'datalang'.  It uses a YAML file as the spec to translate a data set.
}
\examples{
library(datalang)
my_spec <- system.file("specs/thisweek.yml", package = "datalang")
translate_data(my_spec)
}
