% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_folder}
\alias{translate_folder}
\title{Translates and saves multiple a data sets}
\usage{
translate_folder(
  spec_folder = "inst/specs",
  data_folder = "data",
  rd_folder = "man"
)
}
\arguments{
\item{spec_folder}{The path to the folder where the YAML spec files are located. Defaults to 'inst/specs'.}

\item{data_folder}{The target folder location where the 'rda' file will be saved to.}

\item{rd_folder}{The target folder location where the help files will be saved to.}
}
\description{
Cycles through all of the spec files inside a folder specified in the 'spec_folder' argument.
The function translates and saves them to a specified folder. It also creates and saves the
help files.

Saves a translated versions of the data set in 'rda' format.

This function is meant for packages that will ship with a copy of the translated data set.
}
\examples{
library(datalang)
my_spec_folder <- system.file("specs", package = "datalang")
folder_data(my_spec_folder, tempdir())
}
